+++ docs/CHANGELOG.md
@@ -3,6 +3,7 @@
 * Added `is:armorintrinsic` to find Artifice Armor, armor with seasonal perks, etc.
 * Compare suggestion buttons now offer comparison to similar armor intrinsics.
 * Added perks to Light.gg links. See your weapon's popularity rating without having to reselect its perks.
+* Removed the "streak" boxes from Trials rank.
 
 ## 7.38.0 <span class="changelog-date">(2022-10-09)</span>
